refactor: split analysis/subagents/priority.py into focused modules
